+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Removed `verb-max-redirections` variable, as it wasn't working as intended. Use `url-max-redirections` instead.
 - Add `default` argument to `verb-var`.
 - Removed automatic addition of `charset=` value in `Content-Type` request headers.
+- Add optional `coding-system` argument to `verb-read-file`.
 
 ## **2.5.0** - 2020-01-30
 - Added `verb-trim-body-end` variable.