+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Verb Changelog
-## master
+## **2.6.0** - 2020-02-03
 - Removed `verb-max-redirections` variable, as it wasn't working as intended. Use `url-max-redirections` instead.
 - Add `default` argument to `verb-var`.
 - Removed automatic addition of `charset=` value in `Content-Type` request headers.